update consensus logic
